requirements.txt
botbuilder-core==4.16.2
botbuilder-schema==4.16.2
botbuilder-integration-aiohttp==4.16.2
botframework-connector==4.16.2
msal==1.34.0
PyJWT==2.8.0
